2025 06 tf cassandra (#87)
Refactored TableInserter to drain writers in a truly async way. Added a Terraform script for Cassandra cluster deployment.

+++ test/code/tag_and_denormalize/4_clean.sh
@@ -1,8 +1,20 @@
 #!/bin/bash
 
-outDir=/tmp/capi_out/tag_and_denormalize_quicktest
+fs_or_https=$1
+one_or_multi=$2
+
+if [[ "$fs_or_https" != "fs" && "$fs_or_https" != "https" || \
+  "$one_or_multi" != "one" && "$one_or_multi" != "multi" ]]; then
+  echo $(basename "$0") requires 2 parameters: 'fs|https' 'one|multi'
+  exit 1
+fi
+
+dataDirName=tag_and_denormalize_quicktest
+keyspace=${dataDirName}_${fs_or_https}_${one_or_multi}
+
+outDir=/tmp/capi_out/$dataDirName
 
 rm -f $outDir/tag_totals.tsv $outDir/tagged_products_for_operator_review.csv $outDir/runs.csv
 pushd ../../../pkg/exe/toolbelt
-  go run capitoolbelt.go drop_keyspace -keyspace=tag_and_denormalize_quicktest
+  go run capitoolbelt.go drop_keyspace -keyspace=$keyspace
 popd
